requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
djangorestframework==3.15.2
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.1.0rc8
future==1.0.0
idna==3.8
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
pymongo==4.8.0
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
